+++ lib/dynamic_widget/basic/align_widget_parser.dart
@@ -9,7 +9,7 @@ class AlignWidgetParser extends WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     return Align(
       alignment: map.containsKey("alignment")
           ? parseAlignment(map["alignment"])
@@ -17,7 +17,7 @@ class AlignWidgetParser extends WidgetParser {
       widthFactor: map.containsKey("widthFactor") ? map["widthFactor"] : null,
       heightFactor:
           map.containsKey("heightFactor") ? map["heightFactor"] : null,
-      child: DynamicWidgetBuilder.buildFromMap(map["child"], listener),
+      child: DynamicWidgetBuilder.buildFromMap(map["child"], buildContext, listener),
     );
   }
 }