+++ lib/dynamic_widget/basic/cliprrect_widget_parser.dart
@@ -10,7 +10,7 @@ class ClipRRectWidgetParser extends WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     var radius = map['borderRadius'].toString().split(",");
     double topLeft = double.parse(radius[0]);
     double topRight = double.parse(radius[1]);
@@ -24,7 +24,7 @@ class ClipRRectWidgetParser extends WidgetParser {
           bottomLeft: Radius.circular(bottomLeft),
           bottomRight: Radius.circular(bottomRight)),
       clipBehavior: parseClipBehavior(clipBehaviorString),
-      child: DynamicWidgetBuilder.buildFromMap(map["child"], listener),
+      child: DynamicWidgetBuilder.buildFromMap(map["child"], buildContext, listener),
     );
   }
 }