+++ lib/dynamic_widget/basic/container_widget_parser.dart
@@ -9,7 +9,7 @@ class ContainerWidgetParser extends WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     Alignment alignment = parseAlignment(map['alignment']);
     Color color = parseHexColor(map['color']);
     BoxConstraints constraints = parseBoxConstraints(map['constraints']);
@@ -19,7 +19,7 @@ class ContainerWidgetParser extends WidgetParser {
     Map<String, dynamic> childMap = map['child'];
     Widget child = childMap == null
         ? null
-        : DynamicWidgetBuilder.buildFromMap(childMap, listener);
+        : DynamicWidgetBuilder.buildFromMap(childMap, buildContext, listener);
 
     String clickEvent =
     map.containsKey("click_event") ? map['click_event'] : "";