+++ lib/dynamic_widget/basic/fittedbox_widget_parser.dart
@@ -9,13 +9,13 @@ class FittedBoxWidgetParser extends WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     return FittedBox(
       alignment: map.containsKey("alignment")
           ? parseAlignment(map["alignment"])
           : Alignment.center,
       fit: map.containsKey("fit") ? parseBoxFit(map["fit"]) : BoxFit.contain,
-      child: DynamicWidgetBuilder.buildFromMap(map["child"], listener),
+      child: DynamicWidgetBuilder.buildFromMap(map["child"], buildContext, listener),
     );
   }
 }