+++ lib/dynamic_widget/basic/text_widget_parser.dart
@@ -9,7 +9,7 @@ class TextWidgetParser implements WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     String data = map['data'];
     String textAlignString = map['textAlign'];
     String overflow = map['overflow'];