+++ lib/dynamic_widget/basic/wrap_widget_parser.dart
@@ -9,7 +9,7 @@ class WrapWidgetParser extends WidgetParser {
   }
 
   @override
-  Widget parse(Map<String, dynamic> map, ClickListener listener) {
+  Widget parse(Map<String, dynamic> map, BuildContext buildContext, ClickListener listener) {
     return Wrap(
       direction: map.containsKey("direction")
           ? parseAxis(map["direction"])
@@ -31,7 +31,7 @@ class WrapWidgetParser extends WidgetParser {
       verticalDirection: map.containsKey("verticalDirection")
           ? parseVerticalDirection(map["verticalDirection"])
           : VerticalDirection.down,
-      children: DynamicWidgetBuilder.buildWidgets(map['children'], listener),
+      children: DynamicWidgetBuilder.buildWidgets(map['children'], buildContext, listener),
     );
   }
 }